+++ src/client/files.ts
@@ -51,7 +51,7 @@ export async function storeFile(
     sha256 ||
     Array.from(
       new Uint8Array(
-        await crypto.subtle.digest("SHA-256", await blob.arrayBuffer()),
+        await crypto.subtle.digest("SHA-256", await blob.slice().arrayBuffer()),
       ),
     )
       .map((b) => b.toString(16).padStart(2, "0"))